+++ src/eslint-config-adeira/index.js
@@ -1,6 +1,7 @@
 // @flow
 
-const ourRules = require('./ourRules');
+const { ourRules } = require('./ourRules');
+const prettierRules = require('./extraPrettierRules');
 const changeNextVersionErrorLevel = require('./changeNextVersionErrorLevel');
 const getCommonConfig = require('./getCommonConfig');
 
@@ -12,4 +13,9 @@ import type { EslintConfig } from './EslintConfig.flow';
 
 const WARN = 1;
 
-module.exports = (getCommonConfig(changeNextVersionErrorLevel(ourRules, WARN)) /*: EslintConfig */);
+const rules = {
+  ...ourRules,
+  ...prettierRules,
+};
+
+module.exports = (getCommonConfig(changeNextVersionErrorLevel(rules, WARN)) /*: EslintConfig */);