+++ src/eslint-config-adeira/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Unreleased
 
+Config subsets added for use cases where the main default config does not fit well.
+
+- `@adeira/eslint-config/base`
+- `@adeira/eslint-config/jest`
+- `@adeira/eslint-config/react`
+- `@adeira/eslint-config/flowtype`
+- `@adeira/eslint-config/relay`
+
 # 6.1.0
 
 - Rule `import/no-extraneous-dependencies` now ignores `**/__flowtests__/**`, `**/__tests__/**` and `**/.storybook/**` by default.